refactor(agent): clarify session selection storage
Rename the mixed active/boot persistence module to reflect both responsibilities and remove the unused clearActiveSessionId wrapper in favor of writeActiveSessionId(undefined).
