+++ packages/agent/src/front/chat/__tests__/PiChatPanel.test.tsx
@@ -6,7 +6,7 @@ import type { SessionSummary } from '../../../shared/session'
 import { createInitialPiChatState, type PiChatState } from '../pi/piChatReducer'
 import type { RemotePiSession, RemotePiSessionOptions } from '../pi/remotePiSession'
 import { activeSessionStorageKey, scopedComposerStorageKey, type ActiveSessionStorageLike } from '../session'
-import { bootResumeSessionStorageKey } from '../session/activeSessionStorage'
+import { bootResumeSessionStorageKey } from '../session/sessionSelectionStorage'
 import { ComposerContributionProvider } from '../composerContributions'
 import { PiChatPanel as AddressedPiChatPanel, type PiChatPanelProps } from '../PiChatPanel'
 import type { ComposerBlocker } from '../components/ChatNotices'
